[GraphQL] restructuring mutations and queries.
