+++ pkg/model/http/responses.graphqls
@@ -11,3 +11,9 @@ type Response {
 input QuizResponse {
     Responses: [[Int32!]]!
 }
+
+# Requests that might alter the state of data in the database.
+extend type Mutation {
+    # Request to submit responses to a quiz for marking.
+    takeQuiz(quizID: String!, input: QuizResponse!): Float!
+}
